Refactored data classes to hold the date of the values. Added first approach to layout. Added loading state to the UC.

+++ app/src/main/java/com/example/alejandro/bitcoingrapher/network/model/dto/BitcoinDataResponseDto.kt
@@ -10,5 +10,5 @@ data class BitcoinDataResponseDto (
 
 data class BitcoinDataDto(
     val x: Long,
-    val y: Double
+    val y: Float
 )
